+++ circus-train-api/src/test/java/com/hotels/bdp/circustrain/api/metastore/CloseableMetaStoreClientFactoryTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2016-2017 Expedia Inc.
+ * Copyright (C) 2016-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,31 +16,40 @@
 package com.hotels.bdp.circustrain.api.metastore;
 
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
+import org.apache.thrift.TApplicationException;
 import org.apache.thrift.TException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import com.hotels.bdp.circustrain.api.metastore.CloseableMetaStoreClient;
-import com.hotels.bdp.circustrain.api.metastore.CloseableMetaStoreClientFactory;
+import com.hotels.bdp.circustrain.api.metastore.compatibility.HiveMetaStoreClientCompatibility;
 
 @RunWith(MockitoJUnitRunner.class)
 public class CloseableMetaStoreClientFactoryTest {
 
-  @Mock
-  private IMetaStoreClient delegate;
+  private @Mock IMetaStoreClient delegate;
+  private @Mock HiveMetaStoreClientCompatibility compatibility;
 
   @Test
   public void typical() throws TException {
     try (CloseableMetaStoreClient wrapped = CloseableMetaStoreClientFactory.newInstance(delegate)) {
       wrapped.unlock(1L);
     }
-
     verify(delegate).unlock(1L);
     verify(delegate).close();
+  }
+
+  @Test
+  public void compatibility() throws TException {
+    when(delegate.getTable("db", "tbl")).thenThrow(new TApplicationException());
+    try (CloseableMetaStoreClient wrapped = CloseableMetaStoreClientFactory.newInstance(delegate, compatibility)) {
+      wrapped.getTable("db", "tbl");
+    }
+    verify(compatibility).getTable("db", "tbl");
   }
 
 }